Navigation.jl

Navigation.Navigation — Module

Calculate distances, angles, speeds and points for aviation navigation purposes.

All angles are in radians, distances are standard in meters, and speeds are standard in m/s.

These methods are not to be used for operational purposes.

Implemented Functions:

distance

bearing

final_bearing

midpoint

intermediate_point

destination_point

intersection_point

alongtrackdistance

Vground

head_wind

cross_wind

Implemented Types:

Point(ϕ, λ)

Based on source: www.movable-type.co.uk/scripts/latlong.html source: edwilliams.org/avform.htm

Types

Navigation.Point — Type

Point type with latitude ϕ [rad] and longitude λ [rad]

Functions

Navigation.Vground — Method

Vground(Vtas::Float64, Vwind::Float64, ∠wind::Float64, course::Float64)

Return the ground speed Vground [m/s] given the course [rad], airspeed Vtas [m/s], wind speed Vwind[m/s] and wind direction ∠wind [rad]. The speeds need to use the same speed unit ([m/s] or [kts] or [km/h]).

source: edwilliams.org/avform.htm

Navigation.along_track_distance — Function

along_track_distance(pos₁::Point, pos₂::Point, pos₃::Point[,
radius::Float64=Rₑ_m])

The along_track_distance from the start point pos₁ (Point) to the closest point on the great circle path (defined by points pos₁ and pos₂ (Point)) to the point pos₃ (Point). The radius of the earth can also be given.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.along_track_distance — Function

along_track_distance(∠distance₁₃::Float64, cross_track_∠distance::Float64,
[radius::Float64=Rₑ_m])

The along_track_distance from the start point pos₁ (Point) to the closest point on the great circle path (defined by angular distance ∠distance₁₃ [rad] between pos₁ and pos₃ (Point) and the cross track angular distance cross_track_∠distance [rad]) to the point pos₃ (Point). The radius of the earth can also be given.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.bearing — Method

bearing(pos₁::Point, pos₂::Point)

Return the initial bearing [rad] of the great circle line between the positions pos₁ and pos₂ on a sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(pos₁::Point, pos₂::Point, pos₃::Point[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ (Point) to a great circle path defined by the points pos₁ and pos₂ (Point).

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(∠distance₁₃::Float64, bearing₁₂::Float64,
bearing₁₃::Float64[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ to a great circle path. The distance is calculated using the angular distance between points pos₁ and pos₃ (Point), the bearing [rad] between points pos₁ and pos₃ (Point) , the bearing [rad] between points pos₁ and pos₂ (Point), and the radius of the earth [m].

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(pos₁::Point, bearing::Float64, pos₃::Point[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ (Point) to a great circle path defined by the point pos₁ and bearing [rad].

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_wind — Method

cross_wind(Vwind::Float64, ∠wind::Float64, course::Float64)

Return the cross_wind [m/s] component for a given wind speed Vwind [m/s], course [rad], and wind direction ∠wind [rad]. A positive cross-wind component indicates a wind from the right.

source: edwilliams.org/avform.htm

Navigation.destination_point — Function

destination_point(start_pos::Point, distance::Float64, bearing::Float64[, radius::Float64=Rₑ_m])

Given a start_pos (Point), initial bearing rad, distance [m], and the earth radius [m] calculate the destina­tion_point (Point) travelling along a (shortest distance) great circle arc. (The distance must use the same radius as the radius of the earth.)

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.distance — Function

distance(pos₁::Point, pos₂::Point[, radius::Float64=Rₑ_m])

Return the distance in [m] of the great circle line between the positions pos₁ and pos₂ on a sphere, with a given radius, calculated using the haversine formula. The haversine gives also good estimations at short distances.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.final_bearing — Method

final_bearing(pos₁::Point, pos₂::Point)

Return the final_bearing [rad] of the great circle line between the positions pos₁ and pos₂ on a sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.head_wind — Method

head_wind(Vwind::Float64, ∠wind::Float64, course::Float64)

Return the head_wind [m/s] component for a given wind speed Vwind [m/s], course [rad], and wind direction ∠wind [rad]. A tail-wind is a negative head-wind.

source: edwilliams.org/avform.htm

Navigation.intermediate_point — Function

intermediate_point(pos₁::Point, pos₂::Point[, fraction::Float64 = 0.5])

Return the intermediate_point (Point) at any fraction along the great circle path between two points with positions pos₁ and pos₂. The fraction along the great circle route is such that fraction = 0.0 is at pos₁ and fraction 1.0 is at pos₂.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.intersection_point — Method

intersection_point(pos₁::Point, pos₂::Point, bearing₁₃::Float64,
bearing₂₃::Float64)

Return the intersection point pos₃ (Point) of two great circle paths given two start points (Point) pos₁ and pos₂ and two bearings [rad] from pos₁ to pos₃, and from pos₂ to pos₃.

Under certain circumstances the results can be an ∞ or ambiguous solution.

Source: edwilliams.org/avform.htm

Navigation.midpoint — Method

midpoint(pos₁::Point, pos₂::Point)

Return the half-way point midpoint (Point) on the great circle line between the positions pos₁ and pos₂ on a sphere.

Source: www.movable-type.co.uk/scripts/latlong.html