Navigation.jl

Navigation.Navigation — Module

Calculate distances, angles, speeds and points for aviation navigation purposes.

All angles are in radians, distances are standard in meters, and speeds are standard in m/s.

These methods are not to be used for operational purposes.

Implemented Functions:

distance

bearing

final_bearing

midpoint

intermediate_point

destination_point

intersection_point

alongtrackdistance

crosstrackdistance

Vground

head_wind

cross_wind

normalize

closestpointto_poles

Implemented Types:

Point_rad(ϕ, λ)

Point_deg(ϕ, λ)

Implemented constants:

Rₑ_m Radius Earth in [m]

Based on source: www.movable-type.co.uk/scripts/latlong.html source: edwilliams.org/avform.htm

Types

Navigation.Point_rad — Type

Point type with latitude ϕ [rad] and longitude λ [rad]

Navigation.Point_deg — Type

Point_deg type with latitude ϕ [deg] and longitude λ [deg]

Functions

Navigation.Vground — Method

Vground(Vtas::Float64, Vwind::Float64, ∠wind::Float64, course::Float64)

Return the ground speed Vground [m/s] given the course [rad], airspeed Vtas [m/s], wind speed Vwind[m/s] and wind direction ∠wind [rad]. The speeds need to use the same speed unit ([m/s] or [kts] or [km/h]).

source: edwilliams.org/avform.htm

Navigation.along_track_distance — Function

along_track_distance(pos₁::Point, pos₂::Point, pos₃::Point[,
radius::Float64=Rₑ_m])

The along_track_distance from the start point pos₁ (Point) to the closest point on the great circle path (defined by points pos₁ and pos₂ (Point)) to the point pos₃ (Point). The radius of the earth can also be given.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.along_track_distance — Function

along_track_distance(∠distance₁₃::Float64, cross_track_∠distance::Float64,
[radius::Float64=Rₑ_m])

The along_track_distance from the start point pos₁ (Point) to the closest point on the great circle path (defined by angular distance ∠distance₁₃ [rad] between pos₁ and pos₃ (Point) and the cross track angular distance cross_track_∠distance [rad]) to the point pos₃ (Point). The radius of the earth can also be given.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.bearing — Method

bearing(pos₁::Point, pos₂::Point)

Return the initial bearing of the great circle line between the positions pos₁ and pos₂ on a sphere.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.bearing — Method

bearing(section::RouteSection)

Return the initial bearing [rad] of the route section on a sphere.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.closest_point_to_pole — Method

closestpointtopole(point::Pointrad, bearingrad::Float64) closestpointtopole(point::Pointdeg, bearingdeg::Float64)

Given a starting point [Point_rad] and initial bearing [rad] calculate the closest point to the next pole encountered.

Source: https://www.movable-type.co.uk/scripts/latlong.html #TODO Source for longitude

Navigation.cross_track_distance — Function

cross_track_distance(pos₁::Point, pos₂::Point, pos₃::Point[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ (Point) to a great circle path defined by the points pos₁ and pos₂ (Point).

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(pos₁::Point, bearing::Float64, pos₃::Point[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ (Point) to a great circle path defined by the point pos₁ and bearing [rad].

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(∠distance₁₃::Float64, bearing₁₂::Float64,
bearing₁₃::Float64[, radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ to a great circle path. The distance is calculated using the angular distance between points pos₁ and pos₃ (Point), the bearing [rad] between points pos₁ and pos₃ (Point) , the bearing [rad] between points pos₁ and pos₂ (Point), and the radius of the earth [m].

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_track_distance — Function

cross_track_distance(section::RouteSection, pos₃::Point [,
radius::Float64=Rₑ_m])

Return the cross_track_distance [m] from a point pos₃ (Point) to a route section.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.cross_wind — Method

cross_wind(Vwind::Float64, ∠wind::Float64, course::Float64)

Return the cross_wind [m/s] component for a given wind speed Vwind [m/s], course [rad], and wind direction ∠wind [rad]. A positive cross-wind component indicates a wind from the right.

source: edwilliams.org/avform.htm

Navigation.destination_point — Function

destination_point(start_pos::Point, distance::Float64, bearing::Float64[,
 radius::Float64=Rₑ_m])

Given a start_pos (Point), initial bearing (clockwise from North), distance [m], and the earth radius [m] calculate the destina­tion_point (Point) travelling along a (shortest distance) great circle arc. (The distance must use the same radius as the radius of the earth.)

Point = Navigation.Pointdeg or Navigation.Pointrad (if Pointdeg then the initial bearing must also be in degrees, and if Pointrad then the intial bearing must be in radians)

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.distance — Function

distance(pos₁::Point, pos₂::Point[, radius::Float64=Rₑ_m])

Return the distance in [m] of the great circle line between the positions pos₁ and pos₂ on a sphere, with a given radius, calculated using the haversine formula. The haversine gives also good estimations at short distances.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.distance — Function

distance(section::RouteSection [, radius::Float64=Rₑ_m])

Return the distance in [m] of the great circle line on the route section on a sphere, with a given radius, calculated using the haversine formula. The haversine gives also good estimations at short distances.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.final_bearing — Method

final_bearing(pos₁::Point, pos₂::Point)

Return the final_bearing of the great circle line between the positions pos₁ and pos₂ on a sphere.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.final_bearing — Method

final_bearing(section::RouteSection)

Return the final_bearing [rad] of the route section on a sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.head_wind — Method

head_wind(Vwind::Float64, ∠wind::Float64, course::Float64)

Return the head_wind [m/s] component for a given wind speed Vwind [m/s], course [rad], and wind direction ∠wind [rad]. A tail-wind is a negative head-wind.

source: edwilliams.org/avform.htm

Navigation.intermediate_point — Function

intermediate_point(pos₁::Point, pos₂::Point[, fraction::Float64 = 0.5])

Return the intermediate_point (Point) at any fraction along the great circle path between two points with positions pos₁ and pos₂. The fraction along the great circle route is such that fraction = 0.0 is at pos₁ and fraction 1.0 is at pos₂.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.intermediate_point — Function

intermediate_point(section::RouteSection [, fraction::Float64 = 0.5])

Return the intermediate_point (Point) at any fraction along the route section. The fraction along the great circle route is such that fraction = 0.0 is at pos₁ and fraction 1.0 is at pos₂.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.intersection_point — Method

intersection_point(pos₁::Point, pos₂::Point, bearing₁₃::Float64,
bearing₂₃::Float64)

Return the intersection point pos₃ (Point) of two great circle paths given two start points (Point) pos₁ and pos₂ and two bearings from pos₁ to pos₃, and from pos₂ to pos₃.

Point = Navigation.Pointdeg or Navigation.Pointrad (if Pointdeg then the initial bearings must also be in degrees, and if Pointrad then the intial bearings must be in radians)

Under certain circumstances the results can be an ∞ or ambiguous solution.

Source: edwilliams.org/avform.htm

Navigation.max_latitude_deg — Method

maxlatitudedeg(latdeg::Float64, bearingdeg::Float64)

Using Clairaut's formula it is possible to calculate the maximum latitude of a great circle path, give a bearing bearing_deg [deg] and latitude lat_deg [deg] on the great circle.

The minimum latitude is -maxlatitudedeg

Source: https://www.movable-type.co.uk/scripts/latlong.html

Navigation.max_latitude_rad — Method

maxlatituderad(latrad::Float64, bearingrad::Float64)

Using Clairaut's formula it is possible to calculate the maximum latitude of a great circle path, give a bearing bearing_rad [rad] and latitude lat_rad [rad] on the great circle.

The minimum latitude is -maxlatituderad

Source: https://www.movable-type.co.uk/scripts/latlong.html

Navigation.midpoint — Method

midpoint(pos₁::Point, pos₂::Point)

Return the half-way point midpoint (Point) on the great circle line between the positions pos₁ and pos₂ on a sphere.

Point = Navigation.Pointdeg or Navigation.Pointrad

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.midpoint — Method

midpoint(section::RouteSection)

Return the half-way point midpoint (Point) on the route section on a sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

Navigation.opposite_point — Method

oppositepoint(point::Pointrad) oppositepoint(point::Pointdeg)

The point at the opposite site of the Earth.